Create PULL_REQUEST_TEMPLATE.md
Per the best practices doc, I have created a pull request template. The creation of these templates can be viewed on GitHub [here](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository). It is my understanding, that we can create a different template for the policies than for the rest of the repo, etc. I think it is worth exploring to enhance the community experience.

Signed-off-by:  Jeff EF <namdeirf@gmail.com>